backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.3
numpy==1.24.3
biopython==1.81
Faker==24.11.0

boto3==1.26.137
botocore==1.29.137
django-storages==1.14.2
celery==5.3.1
redis==5.0.0
django-celery-beat==2.5.0
django-celery-results==2.5.1

openai>=1.6.1,<2.0.0
langchain==0.0.350
langchain-openai==0.0.2
google-generativeai>=0.3.0
matplotlib==3.8.0
seaborn==0.12.2
scikit-learn==1.3.2
plotly==5.17.0
